Sine Generator
It has no options because I don’t know how exactly to go about changing
things. The generator is meant to accept a number of sine waves from
the user to combine.

Problems with the generator is I don’t know the best parameters so the
output doesn’t look so noisy in most cases.